refactor: Move the store caches into the caches module
